[MOB-4248] Refactor displayer to use a base class
