+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableRenderJsonHandler.java
@@ -3,5 +3,5 @@ package com.iterable.iterableapi;
 import org.json.JSONObject;
 
 public interface IterableRenderJsonHandler {
-    void showMessage(JSONObject payload, IterableHelper.IterableUrlCallback urlCallback);
+    boolean showMessage(JSONObject payload, IterableHelper.IterableUrlCallback urlCallback);
 }
